requirements.txt
django
djangorestframework
djangorestframework-simplejwt
redis
pandas
celery
requests
